pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Root workspace configuration
[tool.poetry]
name = "pyHaasAPI"
version = "2.0.0"
description = "Modern, async-first Python library for HaasOnline API integration"
authors = ["pyHaasAPI Team <support@pyhaasapi.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "pyHaasAPI"},
    {include = "pyHaasAPI_v1"},
    {include = "mcp_server"},
]

# Shared dependencies for all projects
[tool.poetry.dependencies]
python = "^3.11"
# Core shared dependencies
pydantic = "^2.0.0"
requests = "^2.31.0"
loguru = "^0.7.0"
typing-extensions = "^4.7.0"
python-dotenv = "^1.0.0"
# Async dependencies for v2
aiohttp = "^3.9.0"
asyncio = "^3.4.3"
# Legacy dependencies for v1
pandas = "^2.0.0"
numpy = "^1.24.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"

# Optional dependencies for specific projects
mcp = {version = "^1.0.0", optional = true}
fastapi = {version = "^0.104.0", optional = true}
uvicorn = {version = "^0.24.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.14.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.7.0"

# Project-specific extras
[tool.poetry.extras]
pyhaasapi = []  # Core library has no extra deps
mcp-server = ["mcp", "fastapi", "uvicorn"]
all = ["mcp", "fastapi", "uvicorn"]

# Workspace configuration
[tool.poetry.group.pyhaasapi.dependencies]
# pyHaasAPI specific dependencies (if any)

[tool.poetry.group.mcp-server.dependencies]
# MCP server specific dependencies
mcp = "^1.0.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that interact with real API (deselect with '-m \"not integration\"')",
]
filterwarnings = [
    "ignore::pydantic.PydanticDeprecatedSince20",
    "ignore:Valid config keys have changed in V2:UserWarning"
]


# CLI entry points
[tool.poetry.scripts]
haas-cli = "pyHaasAPI.cli.main:main"
haas-lab = "pyHaasAPI.cli.lab_cli:main"
haas-bot = "pyHaasAPI.cli.bot_cli:main"
haas-analysis = "pyHaasAPI.cli.analysis_cli:main"
haas-backtest = "pyHaasAPI.cli.backtest_cli:main"
# Legacy v1 CLI
haas-v1-analyze = "pyHaasAPI_v1.cli.main:main"
